refactor: split functions in utils script
